requirements.txt
agno
groq
duckduckgo-search
yfinance
python-dotenv
ddgs